Create genesis version package, extract state schema interface
