+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/state/beaconstate/BeaconStateInvariants.java
@@ -23,27 +23,34 @@ import tech.pegasys.teku.ssz.backing.schema.SszSchema;
 import tech.pegasys.teku.ssz.backing.view.SszPrimitives;
 import tech.pegasys.teku.ssz.sos.SszField;
 
-class BeaconStateInvariants {
+public class BeaconStateInvariants {
   // Schemas
-  static final SszSchema<SszPrimitives.SszUInt64> GENESIS_TIME_SCHEMA =
+  public static final SszSchema<SszPrimitives.SszUInt64> GENESIS_TIME_SCHEMA =
       SszPrimitiveSchemas.UINT64_SCHEMA;
-  static final SszSchema<SszPrimitives.SszBytes32> GENESIS_VALIDATORS_ROOT_SCHEMA =
+  public static final SszSchema<SszPrimitives.SszBytes32> GENESIS_VALIDATORS_ROOT_SCHEMA =
       SszPrimitiveSchemas.BYTES32_SCHEMA;
-  static final SszSchema<SszPrimitives.SszUInt64> SLOT_SCHEMA = SszPrimitiveSchemas.UINT64_SCHEMA;
+  public static final SszSchema<SszPrimitives.SszUInt64> SLOT_SCHEMA =
+      SszPrimitiveSchemas.UINT64_SCHEMA;
 
   // Fields
-  static SszField GENESIS_TIME_FIELD =
+  public static SszField GENESIS_TIME_FIELD =
       new SszField(0, BeaconStateFields.GENESIS_TIME.name(), GENESIS_TIME_SCHEMA);
-  static SszField GENESIS_VALIDATORS_ROOT_FIELD =
+  public static SszField GENESIS_VALIDATORS_ROOT_FIELD =
       new SszField(
           1, BeaconStateFields.GENESIS_VALIDATORS_ROOT.name(), GENESIS_VALIDATORS_ROOT_SCHEMA);
-  static SszField SLOT_FIELD = new SszField(2, BeaconStateFields.SLOT.name(), SLOT_SCHEMA);
+  public static SszField SLOT_FIELD = new SszField(2, BeaconStateFields.SLOT.name(), SLOT_SCHEMA);
 
   // Return list of invariant fields
   public static List<SszField> getInvariantFields() {
     return List.of(GENESIS_TIME_FIELD, GENESIS_VALIDATORS_ROOT_FIELD, SLOT_FIELD);
   }
 
+  /**
+   * Extract the slot value from any serialized state
+   *
+   * @param bytes A serialized state
+   * @return The slot of the state
+   */
   public static UInt64 extractSlot(final Bytes bytes) {
     // Check assumptions
     checkState(GENESIS_TIME_SCHEMA.isFixedSize(), "Expected genesisTime field to be a fixed size");
