+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/state/beaconstate/BeaconStateSchema.java
@@ -13,18 +13,8 @@
 
 package tech.pegasys.teku.spec.datastructures.state.beaconstate;
 
-import static com.google.common.base.Preconditions.checkArgument;
-import static tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateInvariants.GENESIS_TIME_FIELD;
-import static tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateInvariants.GENESIS_VALIDATORS_ROOT_FIELD;
-import static tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateInvariants.SLOT_FIELD;
-
-import com.google.common.annotations.VisibleForTesting;
-import java.util.List;
-import java.util.stream.Collectors;
-import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.spec.constants.SpecConstants;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlockHeader;
 import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
 import tech.pegasys.teku.spec.datastructures.state.Checkpoint;
@@ -34,213 +24,15 @@ import tech.pegasys.teku.spec.datastructures.state.Validator;
 import tech.pegasys.teku.ssz.SSZTypes.SSZList;
 import tech.pegasys.teku.ssz.SSZTypes.SSZVector;
 import tech.pegasys.teku.ssz.backing.collections.SszBitvector;
-import tech.pegasys.teku.ssz.backing.schema.AbstractSszContainerSchema;
-import tech.pegasys.teku.ssz.backing.schema.SszListSchema;
-import tech.pegasys.teku.ssz.backing.schema.SszPrimitiveSchemas;
-import tech.pegasys.teku.ssz.backing.schema.SszSchemaHints;
-import tech.pegasys.teku.ssz.backing.schema.SszVectorSchema;
-import tech.pegasys.teku.ssz.backing.schema.collections.SszBitvectorSchema;
-import tech.pegasys.teku.ssz.backing.tree.TreeNode;
-import tech.pegasys.teku.ssz.sos.SszField;
+import tech.pegasys.teku.ssz.backing.schema.SszContainerSchema;
 
-public class BeaconStateSchema extends AbstractSszContainerSchema<BeaconState> {
+public interface BeaconStateSchema<T extends BeaconState, TMutable extends MutableBeaconState>
+    extends SszContainerSchema<T> {
+  TMutable createBuilder();
 
-  @VisibleForTesting
-  BeaconStateSchema(final List<SszField> fields) {
-    super(
-        "BeaconState",
-        fields.stream()
-            .map(f -> namedSchema(f.getName(), f.getSchema().get()))
-            .collect(Collectors.toList()));
-    validateFields(fields);
-  }
+  T createEmpty();
 
-  private void validateFields(final List<SszField> fields) {
-    for (int i = 0; i < fields.size(); i++) {
-      final int fieldIndex = fields.get(i).getIndex();
-      checkArgument(
-          fieldIndex == i,
-          "%s fields must be ordered and contiguous.  Encountered unexpected index %s at fields element %s",
-          getClass().getSimpleName(),
-          fieldIndex,
-          i);
-    }
-
-    final List<SszField> invariantFields = BeaconStateInvariants.getInvariantFields();
-    checkArgument(
-        fields.size() >= invariantFields.size(),
-        "Must provide at least %s fields",
-        invariantFields.size());
-    for (SszField invariantField : invariantFields) {
-      final int invariantIndex = invariantField.getIndex();
-      final SszField actualField = fields.get(invariantIndex);
-      checkArgument(
-          actualField.equals(invariantField),
-          "Expected invariant field '%s' at index %s, but got '%s'",
-          invariantField.getName(),
-          invariantIndex,
-          actualField.getName());
-    }
-  }
-
-  /**
-   * Extract the slot value from any serialized state
-   *
-   * @param bytes A serialized state
-   * @return The slot of the state
-   */
-  public static UInt64 extractSlot(final Bytes bytes) {
-    return BeaconStateInvariants.extractSlot(bytes);
-  }
-
-  public static BeaconStateSchema create(final SpecConstants specConstants) {
-    return new BeaconStateSchema(getFields(specConstants));
-  }
-
-  private static List<SszField> getFields(final SpecConstants specConstants) {
-    SszField fork_field = new SszField(3, BeaconStateFields.FORK.name(), Fork.SSZ_SCHEMA);
-    // TODO(#3658) - Use non-static BeaconBlockHeaderSchema
-    final BeaconBlockHeader.BeaconBlockHeaderSchema blockHeaderSchema =
-        BeaconBlockHeader.SSZ_SCHEMA;
-    SszField latestBlockHeaderField =
-        new SszField(4, BeaconStateFields.LATEST_BLOCK_HEADER.name(), blockHeaderSchema);
-    SszField blockRootsField =
-        new SszField(
-            5,
-            BeaconStateFields.BLOCK_ROOTS.name(),
-            () ->
-                SszVectorSchema.create(
-                    SszPrimitiveSchemas.BYTES32_SCHEMA, specConstants.getSlotsPerHistoricalRoot()));
-    SszField stateRootsField =
-        new SszField(
-            6,
-            BeaconStateFields.STATE_ROOTS.name(),
-            () ->
-                SszVectorSchema.create(
-                    SszPrimitiveSchemas.BYTES32_SCHEMA, specConstants.getSlotsPerHistoricalRoot()));
-    SszField historicalRootsField =
-        new SszField(
-            7,
-            BeaconStateFields.HISTORICAL_ROOTS.name(),
-            () ->
-                SszListSchema.create(
-                    SszPrimitiveSchemas.BYTES32_SCHEMA, specConstants.getHistoricalRootsLimit()));
-    SszField eth1DataField =
-        new SszField(8, BeaconStateFields.ETH1_DATA.name(), Eth1Data.SSZ_SCHEMA);
-    SszField eth1DataVotesField =
-        new SszField(
-            9,
-            BeaconStateFields.ETH1_DATA_VOTES.name(),
-            () ->
-                SszListSchema.create(
-                    Eth1Data.SSZ_SCHEMA,
-                    (long) specConstants.getEpochsPerEth1VotingPeriod()
-                        * specConstants.getSlotsPerEpoch()));
-    SszField eth1DepositIndexField =
-        new SszField(
-            10, BeaconStateFields.ETH1_DEPOSIT_INDEX.name(), SszPrimitiveSchemas.UINT64_SCHEMA);
-    SszField validatorsField =
-        new SszField(
-            11,
-            BeaconStateFields.VALIDATORS.name(),
-            () ->
-                SszListSchema.create(
-                    Validator.SSZ_SCHEMA,
-                    specConstants.getValidatorRegistryLimit(),
-                    SszSchemaHints.sszSuperNode(8)));
-    SszField balancesField =
-        new SszField(
-            12,
-            BeaconStateFields.BALANCES.name(),
-            () ->
-                SszListSchema.create(
-                    SszPrimitiveSchemas.UINT64_SCHEMA, specConstants.getValidatorRegistryLimit()));
-    SszField randaoMixesField =
-        new SszField(
-            13,
-            BeaconStateFields.RANDAO_MIXES.name(),
-            () ->
-                SszVectorSchema.create(
-                    SszPrimitiveSchemas.BYTES32_SCHEMA,
-                    specConstants.getEpochsPerHistoricalVector()));
-    SszField slashingsField =
-        new SszField(
-            14,
-            BeaconStateFields.SLASHINGS.name(),
-            () ->
-                SszVectorSchema.create(
-                    SszPrimitiveSchemas.UINT64_SCHEMA,
-                    specConstants.getEpochsPerSlashingsVector()));
-    final SszField previousEpochAttestationsField =
-        new SszField(
-            15,
-            BeaconStateFields.PREVIOUS_EPOCH_ATTESTATIONS.name(),
-            () ->
-                SszListSchema.create(
-                    PendingAttestation.SSZ_SCHEMA,
-                    (long) specConstants.getMaxAttestations() * specConstants.getSlotsPerEpoch()));
-    final SszField currentEpochAttestationsField =
-        new SszField(
-            16,
-            BeaconStateFields.CURRENT_EPOCH_ATTESTATIONS.name(),
-            () ->
-                SszListSchema.create(
-                    PendingAttestation.SSZ_SCHEMA,
-                    (long) specConstants.getMaxAttestations() * specConstants.getSlotsPerEpoch()));
-    SszField justificationBitsField =
-        new SszField(
-            17,
-            BeaconStateFields.JUSTIFICATION_BITS.name(),
-            () -> SszBitvectorSchema.create(specConstants.getJustificationBitsLength()));
-    SszField previousJustifiedCheckpointField =
-        new SszField(
-            18, BeaconStateFields.PREVIOUS_JUSTIFIED_CHECKPOINT.name(), Checkpoint.SSZ_SCHEMA);
-    SszField currentJustifiedCheckpointField =
-        new SszField(
-            19, BeaconStateFields.CURRENT_JUSTIFIED_CHECKPOINT.name(), Checkpoint.SSZ_SCHEMA);
-    SszField finalizedCheckpointField =
-        new SszField(20, BeaconStateFields.FINALIZED_CHECKPOINT.name(), Checkpoint.SSZ_SCHEMA);
-
-    return List.of(
-        GENESIS_TIME_FIELD,
-        GENESIS_VALIDATORS_ROOT_FIELD,
-        SLOT_FIELD,
-        fork_field,
-        latestBlockHeaderField,
-        blockRootsField,
-        stateRootsField,
-        historicalRootsField,
-        eth1DataField,
-        eth1DataVotesField,
-        eth1DepositIndexField,
-        validatorsField,
-        balancesField,
-        randaoMixesField,
-        slashingsField,
-        // TODO(#3658) - Make these fields version-specific
-        previousEpochAttestationsField,
-        currentEpochAttestationsField,
-        // Remaining fields are not yet versioned
-        justificationBitsField,
-        previousJustifiedCheckpointField,
-        currentJustifiedCheckpointField,
-        finalizedCheckpointField);
-  }
-
-  @Override
-  public BeaconState createFromBackingNode(TreeNode node) {
-    return new BeaconStateImpl(this, node);
-  }
-
-  public MutableBeaconState createBuilder() {
-    return new MutableBeaconStateImpl(createEmptyBeaconStateImpl(), true);
-  }
-
-  public BeaconState createEmpty() {
-    return createEmptyBeaconStateImpl();
-  }
-
-  public BeaconState create(
+  T create(
 
       // Versioning
       UInt64 genesis_time,
@@ -277,36 +69,5 @@ public class BeaconStateSchema extends AbstractSszContainerSchema<BeaconState> {
       SszBitvector justification_bits,
       Checkpoint previous_justified_checkpoint,
       Checkpoint current_justified_checkpoint,
-      Checkpoint finalized_checkpoint) {
-
-    return createEmpty()
-        .updated(
-            state -> {
-              state.setGenesis_time(genesis_time);
-              state.setGenesis_validators_root(genesis_validators_root);
-              state.setSlot(slot);
-              state.setFork(fork);
-              state.setLatest_block_header(latest_block_header);
-              state.getBlock_roots().setAll(block_roots);
-              state.getState_roots().setAll(state_roots);
-              state.getHistorical_roots().setAll(historical_roots);
-              state.setEth1_data(eth1_data);
-              state.getEth1_data_votes().setAll(eth1_data_votes);
-              state.setEth1_deposit_index(eth1_deposit_index);
-              state.getValidators().setAll(validators);
-              state.getBalances().setAll(balances);
-              state.getRandao_mixes().setAll(randao_mixes);
-              state.getSlashings().setAll(slashings);
-              state.getPrevious_epoch_attestations().setAll(previous_epoch_attestations);
-              state.getCurrent_epoch_attestations().setAll(current_epoch_attestations);
-              state.setJustification_bits(justification_bits);
-              state.setPrevious_justified_checkpoint(previous_justified_checkpoint);
-              state.setCurrent_justified_checkpoint(current_justified_checkpoint);
-              state.setFinalized_checkpoint(finalized_checkpoint);
-            });
-  }
-
-  private BeaconStateImpl createEmptyBeaconStateImpl() {
-    return new BeaconStateImpl(this);
-  }
+      Checkpoint finalized_checkpoint);
 }
