+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/gossip/encoding/SszSnappyGossipEncodingTest.java
@@ -26,7 +26,9 @@ import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.StatusMessage;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.operations.SignedAggregateAndProof;
+import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateSchema;
+import tech.pegasys.teku.spec.datastructures.state.beaconstate.MutableBeaconState;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 import tech.pegasys.teku.ssz.backing.SszData;
 import tech.pegasys.teku.ssz.backing.schema.SszSchema;
@@ -35,7 +37,7 @@ public class SszSnappyGossipEncodingTest {
   private final Spec spec = SpecFactory.createMinimal();
   private final GossipEncoding encoding = createEncoding();
   private final DataStructureUtil dataStructureUtil = new DataStructureUtil(spec);
-  private final BeaconStateSchema beaconStateSchema =
+  private final BeaconStateSchema<BeaconState, MutableBeaconState> beaconStateSchema =
       spec.getGenesisSchemaDefinitions().getBeaconStateSchema();
 
   private GossipEncoding createEncoding() {
